Get to a complete, if simple visualization
Each LED is a set of frequencies, the intensity is the normalized and
auto-gained volume.  The color just sweeps through hue in an HSV
space.  This is roughly what I had on my sousaphone for honk 2021,
with the exception that there I had 5x the LEDs, so had 5 LEDs for
each group of frequencies and ran them as if they were a VU meter.